netdata: facilitate building netdata with cloud support
This commit adds a `withCloud` option to the netdata package. When
enabled it allows the netdata agent to communicate with the netdata
cloud.

Other than that, we unconditionally wrap the `netdata-claim.sh` script
with a path to openssl and no longer remove the `sbin` output as it is
required by the script.